refactor: merge backend into svelte app
