--- backend/.env.example
@@ -1,6 +1,0 @@
-# Backend environment configuration
-PORT=3001
-# Update with your local or hosted Postgres connection string
-DATABASE_URL=postgresql://postgres:postgres@localhost:5432/course_builder
-# Comma-separated list of allowed origins for CORS
-CORS_ORIGINS=http://localhost:5173,http://127.0.0.1:5173
