--- backend/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "Course builder backend service",
-  "main": "dist/server.js",
-  "type": "module",
-  "scripts": {
-    "dev": "tsx watch src/server.ts",
-    "build": "tsc --project tsconfig.json",
-    "start": "node dist/server.js",
-    "lint": "tsc --noEmit"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "@types/cors": "^2.8.19",
-    "cors": "^2.8.5",
-    "dotenv": "^17.2.2",
-    "express": "^5.1.0",
-    "pg": "^8.16.3"
-  },
-  "devDependencies": {
-    "@types/express": "^5.0.3",
-    "@types/node": "^24.5.2",
-    "@types/pg": "^8.15.5",
-    "tsx": "^4.20.5",
-    "typescript": "^5.9.2"
-  }
-}
