+++ frontend/.env.example
@@ -1,2 +1,2 @@
-# URL to the backend service exposed in docker-compose
-PUBLIC_BACKEND_URL=http://localhost:3001
+# Connection string for Postgres used by the SvelteKit server
+DATABASE_URL=postgresql://postgres:postgres@localhost:5432/course_builder
