+++ frontend/package.json
@@ -11,13 +11,18 @@
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
-	"devDependencies": {
-		"@sveltejs/adapter-auto": "^6.0.0",
-		"@sveltejs/kit": "^2.22.0",
-		"@sveltejs/vite-plugin-svelte": "^6.0.0",
-		"svelte": "^5.0.0",
-		"svelte-check": "^4.0.0",
-		"typescript": "^5.0.0",
-		"vite": "^7.0.4"
-	}
+        "dependencies": {
+                "pg": "^8.16.3"
+        },
+        "devDependencies": {
+                "@sveltejs/adapter-auto": "^6.0.0",
+                "@sveltejs/kit": "^2.22.0",
+                "@sveltejs/vite-plugin-svelte": "^6.0.0",
+                "@types/node": "^22.7.4",
+                "@types/pg": "^8.15.5",
+                "svelte": "^5.0.0",
+                "svelte-check": "^4.0.0",
+                "typescript": "^5.0.0",
+                "vite": "^7.0.4"
+        }
 }
